+++ sdk/inc/azure/core/az_json.h
@@ -73,9 +73,11 @@ typedef struct
 
   /// This read-only field gives access to the slice of the JSON text that represents the token
   /// value, and it shouldn't be modified by the caller.
-  /// If the token straddles non-contiguous buffers, this is set to #AZ_SPAN_NULL.
+  /// If the token straddles non-contiguous buffers, this is set to the partial token value
+  /// available in the last segment.
   /// The user can call #az_json_token_copy_into_span() to get the token value into a contiguous
-  /// buffer. In the case of JSON strings, the slice does not include the surrounding quotes.
+  /// buffer.
+  /// In the case of JSON strings, the slice does not include the surrounding quotes.
   az_span slice;
 
   /// This read-only field gives access to the size of the JSON text slice that represents the token
@@ -86,12 +88,17 @@ typedef struct
 
   struct
   {
+    /// A flag to indicate whether the JSON token straddles more than one buffer segment and is
+    /// split amongst non-contiguous buffers. For tokens created from input JSON payloads within a
+    /// contiguous buffer, this field is always false.
+    bool is_multisegment;
+
     /// A flag to indicate whether the JSON string contained any escaped characters, used as an
     /// optimization to avoid redundant checks. It is meaningless for any other token kind.
     bool string_has_escaped_chars;
 
     /// This is the first segment in the entire JSON payload, if it was non-contiguous. Otherwise,
-    /// its set to #AZ_SPAN_NULL.
+    /// its set to #AZ_SPAN_EMPTY.
     az_span* pointer_to_first_buffer;
 
     /// The segment index within the non-contiguous JSON payload where this token starts.
@@ -385,7 +392,7 @@ az_json_writer_get_bytes_used_in_destination(az_json_writer const* json_writer)
  * @param[in] value The UTF-8 encoded value to be written as a JSON string. The value is escaped
  * before writing.
  *
- * @remarks If \p value is #AZ_SPAN_NULL, the empty JSON string value is written (i.e. "").
+ * @remarks If \p value is #AZ_SPAN_EMPTY, the empty JSON string value is written (i.e. "").
  *
  * @return An #az_result value indicating the result of the operation.
  * @retval #AZ_OK The string value was appended successfully.