+++ sdk/inc/azure/core/az_result.h
@@ -41,31 +41,6 @@ enum
 #define _az_RESULT_MAKE_SUCCESS(facility, code) \
   ((int32_t)(((int32_t)(facility) << 16) | (int32_t)(code)))
 
-/**
- * @brief Convenience macro to return if an operation failed.
- */
-#define AZ_RETURN_IF_FAILED(exp) \
-  do \
-  { \
-    az_result const _result = (exp); \
-    if (az_result_failed(_result)) \
-    { \
-      return _result; \
-    } \
-  } while (0)
-
-/**
- * @brief Convenience macro to return if the provided span is not of the expected, required size.
- */
-#define AZ_RETURN_IF_NOT_ENOUGH_SIZE(span, required_size) \
-  do \
-  { \
-    if (az_span_size(span) < required_size || required_size < 0) \
-    { \
-      return AZ_ERROR_INSUFFICIENT_SPAN_SIZE; \
-    } \
-  } while (0)
-
 // az_result Bits:
 //   - 31 Severity (0 - success, 1 - failure).
 //   - 16..30 Facility.
@@ -153,7 +128,7 @@ typedef enum
   AZ_ERROR_IOT_TOPIC_NO_MATCH = _az_RESULT_MAKE_ERROR(_az_FACILITY_IOT, 1),
 
   /// While iterating, there are no more properties to return.
-  AZ_ERROR_IOT_NO_MORE_PROPERTIES = _az_RESULT_MAKE_ERROR(_az_FACILITY_IOT, 2),
+  AZ_ERROR_IOT_END_OF_PROPERTIES = _az_RESULT_MAKE_ERROR(_az_FACILITY_IOT, 2),
 } az_result;
 
 /**