+++ sdk/samples/iot/pnp/pnp_thermostat_component.c
@@ -7,11 +7,12 @@
 #pragma warning(disable : 4996)
 #endif
 
-#include "iot_sample_common.h"
-#include "sample_pnp.h"
-#include "sample_pnp_mqtt_component.h"
+#include <iot_sample_common.h>
 
-#include "sample_pnp_thermostat_component.h"
+#include "pnp_mqtt_message.h"
+#include "pnp_protocol.h"
+
+#include "pnp_thermostat_component.h"
 
 #include <stdbool.h>
 #include <stddef.h>
@@ -60,21 +61,23 @@ static az_result build_command_response_payload(
 
   // Build the command response payload
   az_json_writer jw;
-  AZ_RETURN_IF_FAILED(az_json_writer_init(&jw, payload, NULL));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_begin_object(&jw));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, report_max_temp_name_span));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_double(
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_init(&jw, payload, NULL));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_begin_object(&jw));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, report_max_temp_name_span));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_double(
       &jw, thermostat_component->max_temperature, DOUBLE_DECIMAL_PLACE_DIGITS));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, report_min_temp_name_span));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_double(
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, report_min_temp_name_span));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_double(
       &jw, thermostat_component->min_temperature, DOUBLE_DECIMAL_PLACE_DIGITS));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, report_avg_temp_name_span));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_double(&jw, avg_temp, DOUBLE_DECIMAL_PLACE_DIGITS));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, report_start_time_name_span));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_string(&jw, start_time_span));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, report_end_time_name_span));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_string(&jw, end_time_span));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_end_object(&jw));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, report_avg_temp_name_span));
+  IOT_SAMPLE_RETURN_IF_FAILED(
+      az_json_writer_append_double(&jw, avg_temp, DOUBLE_DECIMAL_PLACE_DIGITS));
+  IOT_SAMPLE_RETURN_IF_FAILED(
+      az_json_writer_append_property_name(&jw, report_start_time_name_span));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_string(&jw, start_time_span));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, report_end_time_name_span));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_string(&jw, end_time_span));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_end_object(&jw));
 
   *out_payload = az_json_writer_get_bytes_used_in_destination(&jw);
 
@@ -87,12 +90,12 @@ static az_result build_telemetry_message(
     az_span* out_payload)
 {
   az_json_writer jw;
-  AZ_RETURN_IF_FAILED(az_json_writer_init(&jw, payload, NULL));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_begin_object(&jw));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, telemetry_name));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_double(
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_init(&jw, payload, NULL));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_begin_object(&jw));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_property_name(&jw, telemetry_name));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_double(
       &jw, thermostat_component->current_temperature, DOUBLE_DECIMAL_PLACE_DIGITS));
-  AZ_RETURN_IF_FAILED(az_json_writer_append_end_object(&jw));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_writer_append_end_object(&jw));
   *out_payload = az_json_writer_get_bytes_used_in_destination(&jw);
 
   return AZ_OK;
@@ -108,25 +111,26 @@ static az_result invoke_getMaxMinReport(
 {
   // az_result result;
   // Parse the "since" field in the payload.
-  az_span start_time_span = AZ_SPAN_NULL;
+  az_span start_time_span = AZ_SPAN_EMPTY;
   az_json_reader jr;
-  AZ_RETURN_IF_FAILED(az_json_reader_init(&jr, payload, NULL));
-  AZ_RETURN_IF_FAILED(az_json_reader_next_token(&jr));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_reader_init(&jr, payload, NULL));
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_reader_next_token(&jr));
   int32_t incoming_since_value_buffer_len;
-  AZ_RETURN_IF_FAILED(az_json_token_get_string(
+  IOT_SAMPLE_RETURN_IF_FAILED(az_json_token_get_string(
       &jr.token,
       incoming_since_value_buffer,
       sizeof(incoming_since_value_buffer),
       &incoming_since_value_buffer_len));
-  start_time_span
-      = az_span_create((uint8_t*)incoming_since_value_buffer, incoming_since_value_buffer_len);
 
   // Set the response payload to error if the "since" field was not sent
-  if (az_span_ptr(start_time_span) == NULL)
+  if (incoming_since_value_buffer_len == 0)
   {
     response = report_error_payload;
     return AZ_ERROR_ITEM_NOT_FOUND;
   }
+
+  start_time_span
+      = az_span_create((uint8_t*)incoming_since_value_buffer, incoming_since_value_buffer_len);
 
   // Get the current time as a string
   time_t rawtime;
@@ -136,7 +140,7 @@ static az_result invoke_getMaxMinReport(
   size_t len = strftime(end_time_buffer, sizeof(end_time_buffer), iso_spec_time_format, timeinfo);
   az_span end_time_span = az_span_create((uint8_t*)end_time_buffer, (int32_t)len);
 
-  AZ_RETURN_IF_FAILED(build_command_response_payload(
+  IOT_SAMPLE_RETURN_IF_FAILED(build_command_response_payload(
       thermostat_component, start_time_span, end_time_span, response, out_response));
 
   return AZ_OK;