+++ sdk/tests/core/inc/az_test_precondition.h
@@ -31,13 +31,13 @@
 //   separately is advised.
 // - Tests using ASSERT_PRECONDITION_CHECKED(func) currently must not be run in parallel (!).
 
-#define ENABLE_PRECONDITION_CHECK_TESTS() \
-  static jmp_buf g_precond_test_jmp_buf; \
+#define ENABLE_PRECONDITION_CHECK_TESTS()          \
+  static jmp_buf g_precond_test_jmp_buf;           \
   static unsigned int precondition_test_count = 0; \
-  static void az_precondition_test_failed_fn() \
-  { \
-    precondition_test_count++; \
-    longjmp(g_precond_test_jmp_buf, 0); \
+  static void az_precondition_test_failed_fn()     \
+  {                                                \
+    precondition_test_count++;                     \
+    longjmp(g_precond_test_jmp_buf, 0);            \
   }
 
 #define SETUP_PRECONDITION_CHECK_TESTS() \
@@ -47,17 +47,17 @@
 // function parameters not being used. Explicitly storing the function result as a bool and using
 // (void) to cast it away so that we don't get a warning related to unused variables, particularly
 // in release configurations.
-#define ASSERT_PRECONDITION_CHECKED(fn) \
-  do \
-  { \
-    precondition_test_count = 0; \
-    (void)setjmp(g_precond_test_jmp_buf); \
-    if (precondition_test_count == 0) \
-    { \
-      bool const result = (fn); \
-      assert(result); \
-      (void)result; \
-    } \
+#define ASSERT_PRECONDITION_CHECKED(fn)           \
+  do                                              \
+  {                                               \
+    precondition_test_count = 0;                  \
+    (void)setjmp(g_precond_test_jmp_buf);         \
+    if (precondition_test_count == 0)             \
+    {                                             \
+      bool const result = (fn);                   \
+      assert(result);                             \
+      (void)result;                               \
+    }                                             \
     assert_int_equal(1, precondition_test_count); \
   } while (0)